[hack] Assert in hh_collect that a hashtable entry is not half written
Summary:
To avoid concurrent writes, `write_at()` will synchronously stick the
value `1` in a hashtable entry's `addr` field. Then, once we allocate and store
the data, we replace the `1` with the actual address.

By the time that we call `hh_collect`, there should be no workers writing
anything to the hashtable. So we should never see this `1`. However, if a
worker dies or misbehaves, and the main process somehow survives, `hh_collect`
could segfault. This at least will crash the main process.

Reviewed By: dabek, oulgen

Differential Revision: D3915517

fbshipit-source-id: f559e74044b667d6047b5623e48fe8f6d5eb5137