Generalize and extend tools for mapping over type structures
Summary:
Promise.all was implemented as a "tuple map" use type wrapped in a custom fun.
This diff converts that use type to a def type, updates the libdef for
Promise.all, and removes the custom fun.

A very similar approach can be applied to mapping over object types. This diff
provides two object map tools, $ObjMap and $ObjMapi. Each takes two type
parameters: the first must be of an object kind and the latter must be callable.

The difference between "map" and "mapi" is the same as OCaml's `Map.map` and
`Map.mapi` -- i.e., the former has a single parameter, a value type, while the
latter has two parameters, the key and the value type. Since this is a
type-level mapping, "keys" are string literal types and values are types.

Reviewed By: avikchaudhuri

Differential Revision: D3934450

fbshipit-source-id: e3c2a51100ed844fc1254b41c593b17232ba29b2