split ExportDeclaration into multiple variants
Summary:
`ExportDeclaration` has a bunch of invalid states, like having both `specifiers` and `declarations`, or `default = true` and `source` or `specifiers`.

this splits it into two variants, `ExportNamedDeclaration` and `ExportDefaultDeclaration`, which are (more) valid by construction. this removes a bunch of `failwith`s.

this also matches ESTree, and I'll update estree_translator to split them in a follow up.

Reviewed By: avikchaudhuri

Differential Revision: D3935962

fbshipit-source-id: f71d3900852f9c012e4bf575beb826742bb92b21