shared mem opts, part 1
Summary:
Previously, we were not collecting shared memory garbage until too late, since
all our collection was `gentle and that does nothing until the heap size is 2
times the size after init.

This diff makes it so that we do `aggressive collection periodically, when the
server is idle. This collection triggers when the heap size is 1.2 times the
size after init.

The net effect is that we don't overflow the heap after large rebases.

Reviewed By: gabelevi

Differential Revision: D3916897

fbshipit-source-id: f43c83663e9a93e2b4665cac6b62498a09dec6a7