Fix ./tool test runner's source-map-support require
Summary:
`./tool test` needs to use the same `source-map-support` that
`babel-register` uses (yuck, I know). It hardcoded the path, but this doesn't
work when npm rearranges the modules. I'm surprised it lasted this long!

This is a more robust solution. It should always find the `source-map-support`
that `babel-register` is using.

Reviewed By: samwgoldman

Differential Revision: D3947983

fbshipit-source-id: d3cf8f3b737b5426538a31fec07ed927a18f927a