structured reasons
Summary:
Reason strings get shuttled into and out from shared memory. We can reduce the
memory weight by factoring out common reasons descriptions into an ADT.

An important side effect of this change is that the sort order for reasons is
now different. Changing the order can cause constraints to be evaluated in a
different order.

One innocuous effect of this is that some Ops are no longer displayed with
errors. This kind of perturbation is a problem with the way that ops work. I
have made no effort to preserve op reasons in test output.

Unfortunately, not all perturbations are innocuous. If we change the order of
evaluation where one of the constraints is a unification with `any`, we might
mask other errors. I only noticed a handful of such cases.

Lastly, there is a bug in shadow properties which is exposed by changing the
order of reasons. I have a fix up separately and worked around it here by
carefully ordering the reason description variants.

Reviewed By: avikchaudhuri

Differential Revision: D3947938

fbshipit-source-id: fc2d79388158ba1d9619e82e989a95043e117f21