Store exports separately from propmaps on the context
Summary:
As part of the work to add property variance information, I need to change the
shape of `Type.properties`, which is currently `Type.t SMap.t`. Named exports
share the same shape, and are stored in `Context.property_maps`.

It's true that exports are converted to property maps (and vice versa) for the
purposes of ES/CJS module interop, but the two are not the same thing. Indeed,
variance does not apply to named exports.

To make this refactoring a bit safer, I changed both maps to a new map type with
an opaque key type so it's not possible to mix up props and exports.

After this commit, properties and exports still have the same underlying shape
(while being type-incompatible), but follow-on commits will actually change the
shape of properties and leave exports as-is.

Reviewed By: bhosmer

Differential Revision: D3295183

fbshipit-source-id: 880fa77df19a17769b02eff73a50001b67236832