Parse property variance sigils in object type annotations
Summary:
Non-method properties and indexers in object types should be able to be
"positive" or "negative."

Positive properties can be read from, but not written to--i.e., they can appear
in positive positions like `x.p` but not negative positions like `x.p = e`.
Conversely, negative properties can be written to, but not read from.

This diff only adds parsing support, not the typing rules for the above.
Currently property variance annotations are ignored.

For interfaces and declared classes, which use the object type AST node, methods
are already assumed to be positive. To avoid changing existing behavior, this
diff does not permit variance sigils for methods.

Reviewed By: gabelevi

Differential Revision: D3286932

fbshipit-source-id: b54e9676b208b49686245438cd4734f08015df74