Parse property variance sigils in classes
Summary:
Class properties, but not methods, should be able to be "positive" or
"negative."

This diff only adds parsing support, not the typing rules for the above.
Currently property variance annotations are ignored.

Methods are already assumed to be positive. To avoid changing existing behavior,
this diff does not permit variance sigils for methods.

The only sorta tricky part of this is parsing `+*` and `*+`. This diff takes
both to be an unexpected variance sigil error. The generator-ness of the method
is preserved. This is explicitly tested, so please forgive any wonkiness in the
implementation.

Reviewed By: gabelevi

Differential Revision: D3289787

fbshipit-source-id: 4683a54decd0caca4e6b3ee63e6ce20309f97903