property variance
Summary:
Before this diff, object properties were stored as simple maps from names to
types. This diff changes the structure of properties to a richer representation
that supports property variance and separate getter/setter types.

Property variance can be viewed as set of capabilities. An neutral property
can be read from and written to. A positive property can be read from, but not
written to. A negative property can be written to, but not read from.

A neutral property is invariant w.r.t. object subtyping. A positive property is
covariant. A negative property is contravariant. Providing more fine-grained
subtyping of objects should enable many common patterns, previously rejected by
Flow, to be permitted given the correct annotations.

Property variance also applies to indexers. This diff dovetails with
strictionaries to provide a workaround for many cases where
covariance+immutability was assumed, as it can now be enforced.

Reviewed By: avikchaudhuri

Differential Revision: D3808897

fbshipit-source-id: 6d959b1d08559d47b62ee87a6b8a4d82c65ddfd5