support multiple versions of flow
Summary:
adds a version selector to /try, letting you run master + the 5 most recent versions. it defaults to the most recent release, instead of tracking master like it does now.

when travis builds the site, it writes `flow.js` and the libs to a versioned path, which are uploaded to S3. the old versions are kept in S3. I manually backfilled the last 5 tags to S3.

aside: also fixed the website install instructions, and mocked the github API when not running in Travis
Closes https://github.com/facebook/flow/pull/2546

Reviewed By: jeffmo

Differential Revision: D3949681

Pulled By: mroch

fbshipit-source-id: 5da4625ca58276050a5fd70a70e7367873361459