Fix 'r' to record in `./tool test --watch`
Summary:
I have little faith that
https://github.com/evanw/node-source-map-support/pull/149 will get merged any
time soon. That means we have to make do with out-of-date source maps when
tests change.

Luckily, record is the only command that really relies on the source maps. So
we can just spawn a new process and its source maps will be up to date! Hooray!

Reviewed By: samwgoldman

Differential Revision: D3949020

fbshipit-source-id: 10f34f58c76f14fb948c0a3d9265b4f9150ce393