[hack] Add token schema to generator
Summary:
I'm eventually going to want to generate code for tokens as well, so here's the schema.

There are three kinds of tokens: tokens where the user provides the text (like names or variables), tokens where the text is a given, like "+" or "function", and one token with no text: the end of file marker token.

In an upcoming diff I'll use these to generate code.

Reviewed By: dabek

Differential Revision: D3964458

fbshipit-source-id: b8cc597f534980a07639836bcc9a5ee5a8e6052d