Store last edit in localStorage for next load
Summary:
It always mildly irked me that /try doesn't remember the thing I wrote last time, so here we go.

(Note that I made this diff in the GitHub browser editor, so I have not tested it at all yet. I'll do that and follow up with confirmation in a little bit when I get to the office)
Closes https://github.com/facebook/flow/pull/2012

Reviewed By: mroch

Differential Revision: D3977257

Pulled By: jeffmo

fbshipit-source-id: a05c7d70e42555f6ce9e29c58b9604a627035a7b