Avoid single quotes on Windows
When node exec's on Windows, it uses cmd.exe.
Unlike Powershell and a unix shell, cmd.exe doesn't
recognize single quotes as quotes.

I originally added the --no-auto-start to debug, but then
noticed that the package_json_changes test was racing, and
force-recheck was restarting the server that we just killed.
So I left it in.