Avoid single quotes on Windows
Summary:
When node exec's on Windows, it uses cmd.exe.
Unlike Powershell and a unix shell, cmd.exe doesn't
recognize single quotes as quotes.

I originally added the --no-auto-start to debug, but then
noticed that the package_json_changes test was racing, and
force-recheck was restarting the server that we just killed.
So I left it in.

This fixes the tests that were broken on Windows by the recent
addition of force-recheck to the tool tests
Closes https://github.com/facebook/flow/pull/2591

Reviewed By: jeffmo

Differential Revision: D3982036

Pulled By: gabelevi

fbshipit-source-id: 543cfa759a80f64913d426d6c9cfd99c1dfcfb38