Provide callback to merge module type context with environment
Summary: The dumper needs to merge the type context for a particular file/module with the global context so that external type references are available when dumping the file. This cannot be done inside the dumper because the merging logic is defined in the module that invokes the dumper. Consequently the dumper now takes an additional parameter which provides a call back that will merge a context built by the dumper with the global context that was created before the dumper got invoked.

Reviewed By: samwgoldman

Differential Revision: D3979892

fbshipit-source-id: 3d4e2974a3619a0ab680c47b6b1737e6d492aeea