[hack] Go to dead watchman instance on bad file descriptor
Summary:
When watchman starts shutting down after it returns a socket address
to us, we open a channel to that invalid socket address without getting
any errors. But since the channel doesn't have a valid FD behind it, trying
to actually do anything with that channel (write to it, or get the FD of it)
will throw Bad File Descriptor exception. This gracefully handles that.

Reviewed By: oulgen

Differential Revision: D3985194

fbshipit-source-id: e97c15df51ef8806f3b2a8c55e497f0c4305843f