[hack] Added fileInfo parsing basics
Summary:
This diff creates a new binary called hh_file_parse, which does a fileInfo parsing of the given file. The parser itself is very minimal and currently just matches parenthesis and finds words after function or class. We still need to handle a lot of different cases, typedefs, constants, namespaces, and different types of parsing errors.

The main point of the diff is to get a binary out for testing.

Reviewed By: ericlippert

Differential Revision: D3965077

fbshipit-source-id: 5fb86bb157020b649a32f6492e48679ba654a2bf