Seal the JSX attributes object if there is no spread
Summary:
Currently, if you try to use exact objects for react props, you get
errors about flowing inexact objects to exact objects. This diff does NOT fix
that. To fix that, we probably need to fiddle a bunch more with the complicated
react machinery.

However, the `jsx` pragma code is much more simple. So by sealing this object,
we allow the jsx function to expect an exact object type. See the test for an
example.

Reviewed By: samwgoldman

Differential Revision: D4015376

fbshipit-source-id: 9810691c1fe4601f1c0d19e10c5c75b764310e2d