Collect lower bounds when generator has multiple returns
Summary:
A generator function with multiple returns flows multiple typeapps of
Generator<Y,R,N> into the function return tvar. If the return statement's type
is concrete, they would unify with each other when the poly type was
specialized.

This diff ensures the generator's return type param is open with a lower bound
instead of a concrete type. When the poly type is specialized, the lower bounds
are collected and run into uses as one expects.

Promises behave the same way, which is why they do not have this problem.

Fixes https://github.com/facebook/flow/issues/2475

Reviewed By: gabelevi

Differential Revision: D4019375

fbshipit-source-id: 815eb4442108d1ad494ddb3d37c36abc75bedb43