track location of Identifier.name
Summary:
the location of identifier nodes in the AST encompasses their type annotations. we track the location of the annotation, but not the identifier itself.

this diff turns the `string` into a `Loc.t * string`, pointing at the identifier itself.

note that I didn't expose this in estree_translator since this would be a major breaking change from ESTree.

Reviewed By: gabelevi

Differential Revision: D4069133

fbshipit-source-id: 0177a9f0d561c604dc5ce1a24c691bb5d643209b