add ObjProtoT, use it in place of MixedT for Object.prototype
Summary:
we used `MixedT` to represent `Object.prototype` and the root class type. I think this was just a historical shortcut.

a non-strict `LookupT` on `MixedT` happens when an object's prototype chain is checked. unfortunately, it also happens when doing a `GetPropT` directly on a `MixedT`, causing Flow to not error on `if ((x: mixed).prop)`. This should always be an error because `x` could be `undefined`.

this diff adds a new `ObjProtoT` type (parallels `FunProtoT`). now `MixedT ~> LookupT` is an error, even when non-strict.

Reviewed By: samwgoldman

Differential Revision: D4080892

fbshipit-source-id: 6a2cbbf292a800e902d0075ed6bd2a1052eb0914