Internal AST node for identifier only needs loc & name
Summary:
This only affects our internal AST representation. The external representation stays the same.

Not all identifiers can have a type annotation or can be optional. So it's kind of silly that every node has these properties. Really, these should be properties on the identifier pattern node.

This is a pretty massive rebase, but is mostly just mechanical refactoring directed by the ocaml compiler. The interesting parts are

* The changes to the AST in spider_monkey_ast.ml
* The changes to estree_translator.ml to keep the external AST the same
Closes https://github.com/facebook/flow/pull/2689

Reviewed By: samwgoldman

Differential Revision: D4085699

Pulled By: gabelevi

fbshipit-source-id: 726ab7d25de2f2330ce39c494e801d27b6bf8664