Fix prev/next links in docs pages.
Summary:
https://flowtype.org/docs/

This PR fixes the links on the bottom of three docs pages to match the order implied in the Table of Contents.

Here's how the docs currently flow (if they are to match TOC on left of the page). Below are what they should be (and what the PR tries to implement).

"quickreference" -> "builtins"
**should be**
"quickreference" -> "syntax"

"syntax" -> "arrays"
**should be**
"syntax" -> "builtins"
Closes https://github.com/facebook/flow/pull/2700

Differential Revision: D4095622

Pulled By: gabelevi

fbshipit-source-id: f98939160315688420136311c4a34f905febcbc8