Document the existence of compiled-to-JS Flow parser on npm
I figured this could be useful to promote more prominently on the main
README. Feel free to refine or reject.

Also, while setting up a build for the JS parser, I found I also needed
to install ocamlfind (used in the Makefile) which didn't come alongside
any of the other packages. If this is the case, I can add that to the
parser README.