Improve parse error for unsupported optional class properties
Summary:
Before this diff, if you wrote `class C {+p?:T}`, we would emit an error on the
variance sigil and the question mark (and then more errors because we're in a
bad state).

This diff doesn't add support for optional class properties, which we should do,
but rather just cleans up the error messages to point only to the question mark
itself and not the variance annotation.

Reviewed By: jeffmo

Differential Revision: D4098640

fbshipit-source-id: a3f3893f96ace25d94c0e04736e4d459fa615428