[hack] Type checker support for multiple concrete `as` constraints
Summary:
It is already possible to specify multiple `as` constraints on generic parameters, using either the syntax
```
function foo<T as I as J>(...)
```
on the generic parameters themselves, or `where` constraints such as
```
function foo(...) where T as I, T as J
```
Subtype tests support this feature, so a value of type `T` can be passed to a function expecting an `I` or a `J`. But method invocation (and similar features) is not supported.

There is one design question that hasn't been completely resolved with this diff: what if more than one constraint type is valid? We don't reject overlapping constraints (e.g. and `I` or `J` that both implement the same named method) even though the constraint may not be satisfiable (because the methods have incompatible signatures, for instance). Instead, for now, we reject the ambiguity at the point of invocation. We may wish to revisit this choice in future.

The method `obj_get_` has been refactored a little in this diff, separating off a function `obj_get_concrete_ty` that deals with the case of a concrete class type. Also some renaming of parameters etc. to aid comprehension.

Differential Revision: D4029912

fbshipit-source-id: 9237b8cbd260adcfcdb4395d820c999ad8350c74