fix refinement merging after if statements
Summary:
if a refinement errors, like `let x:mixed; if (x.foo)`, then the tvar that is installed as the `x.foo` refinement never gets a lower bound. the refinement gets merged into the parent env, and since it's unbound, allows `x.foo` to be anything. `(x.foo: string & number)` passes, for example.

this changes `if` to ignore the refinements done in the branches if neither branch exits abnormally. the union of the refinements from all of the branches is the same as the original type. so instead of `if (typeof x === 'string') {}` generating a refinement for "x is a string" and "x is not a string" (for the invisible else branch) and then unioning them, just forget the refinement because "x is a string or not a string" is the same as "we didn't learn anything about x".

Reviewed By: avikchaudhuri

Differential Revision: D4086320

fbshipit-source-id: ba99a6ab8ec6dfc8b76df4bf5002e0f6eecae1d4