ignore errors from internal operations
Summary:
env merging/widening and refinements are two times where we flow types in redundant ways.

for example, when doing a refinement of `T`, we are always flowing some type `U <: T` back into `T`, which never errors.

or, say that we do something like

```
let x: string = 'foo';
while (false) {
  x = 123;
}
```

we flow `number` into `string` on line 3, and error. then we merge the loop's env (where `x: number`) with the outer env (where `x: string`) and error again. but this is always a redundant error. we need to still do it to widen the type of `x` if it's unannotated, but shouldn't error again.

Reviewed By: gabelevi

Differential Revision: D4068713

fbshipit-source-id: 0d63297a688104ec68e5d8d6a3273c735a36ef18