Document the existence of compiled-to-JS Flow parser on npm
Summary:
I figured this could be useful to promote more prominently on the main
README. Feel free to refine or reject.

Also, while setting up a build for the JS parser, I found I also needed
to install ocamlfind (used in the Makefile) which didn't come alongside
any of the other packages. If this is the case, I can add that to the
parser README.
Closes https://github.com/facebook/flow/pull/2707

Differential Revision: D4111230

Pulled By: gabelevi

fbshipit-source-id: e08852eef7c7975311d19eb111fc4c87c0462dc5