fix location of `this` and `super`
Summary:
repositions `this` and `super` to the place they're used instead of having no location (for the global object) or the function/class itself.

aside from making the kinds of errors in the tests better, this is necessary for one-location errors involving `this`, since otherwise it would have no location.

Reviewed By: gabelevi

Differential Revision: D4116699

fbshipit-source-id: fc8d3a96c770cdae1760b29f4aac13f1aa14d15f