improve arithemtic operand errors
Summary:
given an arithemtic operation like `x * y` where `x` is not numeric (a number or Date object), we currently error saying that `x` is incompatible with the whole `x * y` expression, which is a number.

this doesn't need to be a two-sided error, and the message could be better. instead, just say that operands must be numbers.

Reviewed By: gabelevi

Differential Revision: D4113927

fbshipit-source-id: f7b78faff158174d178cb5554560641c9a391540