Let and const variables should go from undeclared -> declared *after* evaluating the initializer expression
Summary:
From the tests and comment, it seems this was done intentionally before -- but AFAICT that was just erroneous:

It is a TDZ error to reference a `const` or `let` binding before the actual declaration statement is evaluated. Additionally, [initializer expressions are evaluated *before* initializing the binding with the resulting value](https://tc39.github.io/ecma262/#sec-let-and-const-declarations-runtime-semantics-evaluation).

As such, it was incorrect to change the state of a binding from undeclared -> declared before evaluating the initializer expression.

Fixing this disallows things like:

```
const x = x;
```

Fixes #1925

Reviewed By: samwgoldman

Differential Revision: D4126129

fbshipit-source-id: e7d57d74af42286845aba54ce21a4e30391e6f6e