[hack] Implement the state traversal algorithm and output line-broken statement
Summary:
Part two:
Take part 1 and do something with it, see details on the algorithm here: http://journal.stuffwithstuff.com/2015/09/08/the-hardest-program-ive-ever-written/
but the basic flow is, take the initial state generated by the builder with no bound rules, and then add it to the state_queue.
While the state_queue is not empty and we haven't found a configuration without overflow, expand_state.
Iterates over all unbound rules and all values of those rules and creates a new state for each, they are then added to state_queue ordered by cost.
When an appropriate state is found, return it and print it.

There is a ton of performance optimizations that can be done here, and a lot of details (no indention yet) are missing but it's a start.

Differential Revision: D3928144

fbshipit-source-id: f05ead727de66e5ef1499a283623829c3dfce825