[hack] Give up on incrementally udpated error list
Summary:
Trying to keep the lazy check time O(C*things that changed + c*dependencies of things that changed) (for c << C) I implemented a way to update error list incrementally. It is based on the fact that "if I rechecked file A and generated following errors in A, then I can replace the errors of A with them".

Now I understand that this is incorrect: it's possibly to do this and generate only part of the errors for A. There can be more errors - some coming from phases of A I didn't run (I might have run only typing, while there were parsing errors), some coming from completely different files (checking B can add errors in A).

Doing it right would require not storing a map of errors from file to the errors it has, but a map from file and phase to the errors it generated. So basically changing failed_X from sets of files to maps from files to error lists.

The way current system avoids this problem, is by always rechecking files that generated any errors, and this what I revert to this diff too. This means that if you got into a place where you have tens of thousands of errors, our responsiveness will suffer, but I'm not convinced if complexity of tracking it correctly is worth handling this edge case for now.

Too cheer myself up: I tried to do the same thing (and made same mistake) that the creator of Hack himself tried (D1387289) and failed (D1401096).

Reviewed By: jopamer

Differential Revision: D4129071

fbshipit-source-id: d186f132742dabf929aed4ab72b93f69967c76d4