Only compare type args for type apps with same root
Summary:
When a type app flows into another type app, it can be expensive to unfold and
structurally compare the resulting specialized types. Furthermore, these
constraints can participate in loops causing lots of unnecessary repeated work
which isn't currently detected by the type app expansion mechanism.

This diff adds a special case for these constraints specifically when the roots
of the type apps are the same. That is, given a constraint A<T> ~> B<U>, if we
know that A = B, we can skip specialization and only generate constraints for
the type args.

For the type args, we still need to be sensitive to type parameter variance, so
this diff introduces a new use type that waits for the root to become concrete
before generating those constraints.

Because we don't specialize and unfold the root, we must generate constraints
for all type args, regardless of where they appear in the resulting type. That
is, previously this constraint was structural and now it has a more nominal
flavor.

Partially fixes https://github.com/facebook/flow/issues/2724. Still takes about
a minute on my machine, more follow-up work necessary.

Reviewed By: avikchaudhuri

Differential Revision: D4134270

fbshipit-source-id: eecaa28562d54c3f3efd091fe9ac1c3e926bfcf1