fix reason when arrays are used as objects
Summary:
when we do object-like things to an ArrT, we look up the Array class from the libs and do the object-like things on that.

we have always (further back than blame goes) used the reason of the upper bound when dealing with an ArrT, but the reason of the lower bound for StrT, BoolT, NumT, FunT, etc.

so, this changes ArrT to match the others. AFAICT, no errors change location. some ops might change because changing the reason changes the order that constraints are resolved, but ops are just fragile that way.

Reviewed By: samwgoldman

Differential Revision: D4163131

fbshipit-source-id: 961fcf8f83b5dbaebbc2919386c261fbd8acfdc6