add missing printer for ExistsT
Summary:
`dump-types` on a file with an ExistsT, e.g. `Foo<*>`, fataled because
there was no type printer for it. wildcards are the devil.

when there's no wildcard, the fallthrough arg no longer makes sense. also since
we need printers for the primitives and `string_of_t` is the only real caller of
`type_printer`, I inlined the "override" from `string_of_t` into `type_printer`.

Reviewed By: gabelevi

Differential Revision: D4169226

fbshipit-source-id: bb5ea6d97c15c94c4062cb209478d88df9a5d013