Lock around GC collects to inform client to keep waiting for server response
Summary:
Currently, if the server is in the middle of GC-ing shared memory, the client has no way to know and just assumes the server is being non-responsive.

This creates a `SharedMem_js` module which includes `SharedMem` but wraps a lock around `collect` in order to communicate to the client when GC is in progress in the server. From there, the client now waits for the GC lock to free rather than just assuming the server is being unresponsive.

Reviewed By: gabelevi

Differential Revision: D4125512

fbshipit-source-id: 6b59d477767a9a608948a7ad11b340929e05fac9