Look up the prototype chain for computed property accesses
Summary:
Before this diff, computed property access for objects was handled by ElemT
exclusively. This diff generalizes GetProp, SetProp, LookupT, and related use
types to handle both named properties and computed element access.

This refactoring unlocks a few things:
* We can properly implement MethodT for `o[p]()`, where `p` is a literal and
  `o[p]` resolves to a function that references `this`.
* We can convert number literals to strings, if a named property matches
  anywhere along the prototype chain.
* Many uses of symbols will require that we traverse the prototype chain.

Note that none of the above are included in this diff, but rather slated for
follow-up work.

Note the invariant that a computed prop ref's type must be concrete (not open)
and will never be a string literal. This invariant holds because we wait for the
indexer element ot resolve with ElemT before generating the lookup constraints,
and string literals are converted in the named property lookups.

This diff also makes a small change to LookupT. Before, the reason for that use
type was the reason for the property. After this change, that reason is
associated with the "prop ref" instead, and the lookup's own reason should be
the "op reason" -- the reason why we are doing a lookup in the first place.

Ref https://github.com/facebook/flow/pull/2293

Reviewed By: gabelevi

Differential Revision: D4149126

fbshipit-source-id: 54fe4f621ebc8e95ac801e1580c158b00555bbca