Support method calls through computed property access
Summary:
Before this diff, method calls of the form o[p]() did not correctly handle
`this`. The `o[p]` part was evaluated, then called as a function -- that is,
`this` was assumed to be the global object.

This diff adds the necessary machinery to evalute the key, then continue on to
call the method with the correct `this`. Specifically, it (1) introduces a
CallElemT, which internally waits on the key using ElemT and (2) extends the
Read/Write variant previous used in ElemT to handle all three actions that we
now support.

Reviewed By: mroch

Differential Revision: D4152186

fbshipit-source-id: 6a6e708f5014a433098ec92fe5af4941917292b5